AccessKit: do not count 'image' as ID
for some reason a lot of images on Tumblr have this explicitly set
(maybe by mobile apps? i wasn't able to reproduce on desktop)
and since it's server-side the localised string is not used/useful.
this probably doesn't do anything for images posted from clients with
the quirk which are not set to English but hopefully i'm wrong

see #390